Criado service, exception personalizada.

+++ src/main/java/desafio_xp_arq_product/v1/controller/TotalProductsResponse.java
@@ -1,4 +1,4 @@
 package desafio_xp_arq_product.v1.controller;
 
-public record TotalProductsResponse(Integer totalProdutos) {
+public record TotalProductsResponse(long totalProdutos) {
 }
